+++ nav2_behaviors/include/nav2_recoveries/behavior.hpp
@@ -29,7 +29,7 @@
 #include "geometry_msgs/msg/twist.hpp"
 #include "nav2_util/simple_action_server.hpp"
 #include "nav2_util/robot_utils.hpp"
-#include "nav2_core/recovery.hpp"
+#include "nav2_core/behavior.hpp"
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wpedantic"
 #include "tf2/utils.h"
@@ -52,7 +52,7 @@ using namespace std::chrono_literals;  //NOLINT
  * @brief An action server recovery base class implementing the action server and basic factory.
  */
 template<typename ActionT>
-class Recovery : public nav2_core::Recovery
+class Recovery : public nav2_core::Behavior
 {
 public:
   using ActionServer = nav2_util::SimpleActionServer<ActionT>;
