+++ nav2_behaviors/include/nav2_recoveries/simple_behavior.hpp
@@ -1,5 +1,4 @@
-// Copyright (c) 2018 Intel Corporation
-// Copyright (c) 2019 Samsung Research America
+// Copyright (c) 2022 Joshua Wallace
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,21 +10,5 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License. Reserved.
+// limitations under the License.
 
-#include <string>
-#include <memory>
-
-#include "nav2_recoveries/behavior_server.hpp"
-#include "rclcpp/rclcpp.hpp"
-
-int main(int argc, char ** argv)
-{
-  rclcpp::init(argc, argv);
-  auto recoveries_node = std::make_shared<recovery_server::BehaviorServer>();
-
-  rclcpp::spin(recoveries_node->get_node_base_interface());
-  rclcpp::shutdown();
-
-  return 0;
-}
