+++ nav2_behaviors/plugins/spin.cpp
@@ -191,4 +191,4 @@ bool Spin::isCollisionFree(
 }  // namespace nav2_recoveries
 
 #include "pluginlib/class_list_macros.hpp"
-PLUGINLIB_EXPORT_CLASS(nav2_recoveries::Spin, nav2_core::Recovery)
+PLUGINLIB_EXPORT_CLASS(nav2_recoveries::Spin, nav2_core::Behavior)
