+++ nav2_behaviors/plugins/spin.hpp
@@ -12,38 +12,45 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_RECOVERIES__PLUGINS__BACK_UP_HPP_
-#define NAV2_RECOVERIES__PLUGINS__BACK_UP_HPP_
+#ifndef NAV2_RECOVERIES__PLUGINS__SPIN_HPP_
+#define NAV2_RECOVERIES__PLUGINS__SPIN_HPP_
 
 #include <chrono>
+#include <string>
 #include <memory>
 
-#include "nav2_recoveries/recovery.hpp"
-#include "nav2_msgs/action/back_up.hpp"
+#include "nav2_recoveries/behavior.hpp"
+#include "nav2_msgs/action/spin.hpp"
+#include "geometry_msgs/msg/quaternion.hpp"
 
 namespace nav2_recoveries
 {
-using BackUpAction = nav2_msgs::action::BackUp;
+using SpinAction = nav2_msgs::action::Spin;
 
 /**
- * @class nav2_recoveries::BackUp
+ * @class nav2_recoveries::Spin
  * @brief An action server recovery for spinning in
  */
-class BackUp : public Recovery<BackUpAction>
+class Spin : public Recovery<SpinAction>
 {
 public:
   /**
-   * @brief A constructor for nav2_recoveries::BackUp
+   * @brief A constructor for nav2_recoveries::Spin
    */
-  BackUp();
-  ~BackUp();
+  Spin();
+  ~Spin();
 
   /**
    * @brief Initialization to run behavior
    * @param command Goal to execute
    * @return Status of recovery
    */
-  Status onRun(const std::shared_ptr<const BackUpAction::Goal> command) override;
+  Status onRun(const std::shared_ptr<const SpinAction::Goal> command) override;
+
+  /**
+   * @brief Configuration of recovery action
+   */
+  void onConfigure() override;
 
   /**
    * @brief Loop function to run behavior
@@ -64,25 +71,19 @@ protected:
     geometry_msgs::msg::Twist * cmd_vel,
     geometry_msgs::msg::Pose2D & pose2d);
 
-  /**
-   * @brief Configuration of recovery action
-   */
-  void onConfigure() override;
+  SpinAction::Feedback::SharedPtr feedback_;
 
-  double min_linear_vel_;
-  double max_linear_vel_;
-  double linear_acc_lim_;
-
-  geometry_msgs::msg::PoseStamped initial_pose_;
-  double command_x_;
-  double command_speed_;
+  double min_rotational_vel_;
+  double max_rotational_vel_;
+  double rotational_acc_lim_;
+  double cmd_yaw_;
+  double prev_yaw_;
+  double relative_yaw_;
+  double simulate_ahead_time_;
   rclcpp::Duration command_time_allowance_{0, 0};
   rclcpp::Time end_time_;
-  double simulate_ahead_time_;
-
-  BackUpAction::Feedback::SharedPtr feedback_;
 };
 
 }  // namespace nav2_recoveries
 
-#endif  // NAV2_RECOVERIES__PLUGINS__BACK_UP_HPP_
+#endif  // NAV2_RECOVERIES__PLUGINS__SPIN_HPP_
