+++ nav2_behaviors/plugins/wait.cpp
@@ -57,4 +57,4 @@ Status Wait::onCycleUpdate()
 }  // namespace nav2_recoveries
 
 #include "pluginlib/class_list_macros.hpp"
-PLUGINLIB_EXPORT_CLASS(nav2_recoveries::Wait, nav2_core::Recovery)
+PLUGINLIB_EXPORT_CLASS(nav2_recoveries::Wait, nav2_core::Behavior)
