+++ nav2_bringup/launch/navigation_launch.py
@@ -173,7 +173,7 @@ def generate_launch_description():
                 remappings=remappings),
             ComposableNode(
                 package='nav2_recoveries',
-                plugin='recovery_server::BehaviorServer',
+                plugin='behavior_server::BehaviorServer',
                 name='recoveries_server',
                 parameters=[configured_params],
                 remappings=remappings),
